requirements.txt
# Core Django
Django
asgiref
sqlparse
tzdata

# Django REST Framework
djangorestframework
django-cors-headers
django-filter

# Authentication & Authorization
dj-rest-auth
django-allauth
djangorestframework_simplejwt
PyJWT

# Google OAuth
google-auth
google-auth-oauthlib
google-auth-httplib2

# Translation Support
django-parler
django-parler-rest
# Celery (Task Queue)
celery
django-celery-beat
redis
kombu
amqp
billiard
click
vine
# Payment Processing
stripe

# AI Functionality
openai
langdetect

# HTTP Clients & Networking
requests
httpx
httpcore
h11
urllib3
certifi
charset-normalizer
idna
anyio

# Image Processing
pillow

# Database
psycopg2-binary

# Web Server
gunicorn

# Data Validation
pydantic
pydantic_core
annotated-types
jiter

# Cryptography
cryptography
cffi
pycparser

# Utilities
python-dotenv
python-dateutil
six
packaging
typing_extensions
firebase-admin